feat(client): Require users to add a user agent
This is something I've been asked to do, and is out
of my control. Whatever happens when the Canister API
recieves the request and returns the requested data,
it's out of my control.

While this is something that this API wrapper needs
to implement, it also improves the Canister API in
one way or another.

Signed-off-by: TheRealKeto <therealketo@gmail.com>